+++ framework/changelog.md
@@ -1,4 +1,4 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- upgrade: core upgrades to 1.1.38
+- upgrade: core upgrades to 1.1.39