+++ transports/changelog.md
@@ -1,5 +1,7 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- Fix: Adds `Base URL` input in custom provider creation dialog.
-- Fix: Fixes `x` button getting hidden behind dialog header.
+- Feat: Adds support for text completion endpoint for OpenAI, SGL.
+- Feat: Adds support for text completion endpoint in OpenAI integration.
+- upgrade: core to 1.1.39
+- upgrade: framework to 1.0.25